Adding toolbar with enter/exit animations to nested preference screens
