+++ WordPress/src/main/res/values/styles.xml
@@ -321,4 +321,10 @@
         <item name="android:background">@drawable/fab_menu_label_background</item>
         <item name="android:textColor">@color/white</item>
     </style>
+
+    <!-- Used in Notifications and Site settings to animate nested preference screens -->
+    <style name="DialogAnimations">
+        <item name="android:windowEnterAnimation">@anim/activity_slide_in_from_right</item>
+        <item name="android:windowExitAnimation">@anim/activity_slide_out_to_right</item>
+    </style>
 </resources>
